requirements.txt
Django == 3.0.5
beautifulsoup4 == 4.9.0
boto3 == 1.12.43
celery == 4.4.2
flower == 0.9.4
python-twitter == 3.5
python-dotenv == 0.13.0
requests == 2.23.0
urllib3 == 1.25.9
psycopg2 == 2.8.5
